pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "polymarket-bot"
version = "0.1.0"
description = "Reward-aware passive Polymarket market maker lane with parity and latency overlay."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.7",
    "pydantic-settings>=2.2",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "requests>=2.31",
    "sqlalchemy>=2.0",
    "psycopg[binary]>=3.2",
    "typer>=0.12",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]
clob = [
    "py-clob-client>=0.34.6",
    "py-builder-relayer-client>=0.0.1",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src", "src.*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
